Completed week02 tasks

+++ Week02/PolarityOfNumber.java
@@ -4,7 +4,7 @@ public class PolarityOfNumber {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.println("Enter the number: ");
-        var a= input.nextFloat();
+        var a= input.nextDouble();
         if(a>0) System.out.println("It is positive!");
         if(a<0) System.out.println("It is neegative!");
         if(a==0) System.out.println("It is neutral!");
